requirements.txt
absl-py==2.1.0
accelerate==0.26.1
antlr4-python3-runtime==4.9.3
array_record==0.5.1
asttokens==3.0.0
astunparse==1.6.3
attrs==23.2.0
bddl==1.0.1
beautifulsoup4==4.13.3
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
chex==0.1.85
click==8.1.7
cloudpickle==2.1.0
contextlib2==21.6.0
cycler==0.12.1
Cython==0.29.37
d4rl==1.1
decorator==4.4.2
decord==0.6.0
distlib==0.3.8
distrax==0.1.5
dm-env==1.6
dm-tree==0.1.8
dm_control==1.0.27
docker-pycreds==0.4.0
easydict==1.9
egl_probe==1.0.2
einops==0.4.1
etils==1.5.2
evdev==1.7.1
exceptiongroup==1.2.1
executing==2.2.0
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.14.0
flatbuffers==24.3.25
flax==0.7.5
fonttools==4.53.0
fsspec==2024.5.0
future==0.18.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
glfw==2.7.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.64.0
gym==0.23.1
gym-notices==0.0.8
h5py==3.11.0
huggingface-hub==0.23.2
hydra-core==1.2.0
hydra-submitit-launcher==1.2.0
identify==2.5.36
idna==3.7
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.18.1
jax==0.4.27
jaxlib==0.4.27
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupytext==1.16.2
keras==2.15.0
kiwisolver==1.4.5
labmaze==1.0.6
libclang==18.1.1
llvmlite==0.42.0
lxml==5.3.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
ml-dtypes==0.2.0
ml_collections==0.1.1
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
mujoco==3.2.7
mujoco-py==2.1.2.14
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.9.0
numba==0.59.1
numpy 
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
nvitop==1.4.2
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.6.0.66
OpenEXR==3.2.4
opt-einsum==3.3.0
optax==0.1.5
orbax-checkpoint==0.4.4
packaging==24.0
parso==0.8.4
pathtools==0.1.2
pexpect==4.9.0
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
pre-commit==3.3.3
proglog==0.1.10
promise==2.3
prompt_toolkit==3.0.50
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.6
pybind11_global==2.13.6
pybullet==3.2.7
pycocotools==2.0.7
pycparser==2.22
pygame==2.5.2
Pygments==2.18.0
pymunk==6.8.0
pynput==1.7.7
PyOpenGL==3.1.7
pyparsing==3.1.2
pytest==8.2.1
python-dateutil==2.9.0.post0
python-xlib==0.33
PyWavelets==1.6.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
robomimic==0.2.0
robosuite==1.4.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.19.3
scikit-learn==1.5.0
scipy==1.13.1
sentry-sdk==2.3.1
setproctitle==1.3.3
shapely==2.0.3
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
soupsieve==2.6
stack-data==0.6.3
submitit==1.5.1
sympy==1.12.1
tenacity==8.3.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.15.0
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.15.0
tensorflow-graphics==2021.12.3
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-metadata==1.15.0
tensorflow-probability==0.23.0
tensorflow-text==2.15.0
tensorstore==0.1.45
termcolor==2.4.0
tf_keras==2.15.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.5.22
timm==1.0.15
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
traitlets==5.14.3
transformers==4.21.1
trimesh==4.4.0
triton==2.3.0
typeguard==2.13.3
typing_extensions==4.12.1
urllib3==2.2.1
virtualenv==20.26.2
wandb==0.13.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.14.1
zipp==3.19.1
